Splits main.rs into main.rs and lib.rs.
